+++ src/cli/subcommands/state_cmd.rs
@@ -3,15 +3,12 @@
 
 use std::path::PathBuf;
 
-use crate::lotus_json::LotusJson;
-use crate::rpc_client::state_ops::state_fetch_root;
+use crate::rpc_client::ApiInfo;
 use crate::shim::clock::ChainEpoch;
 use crate::shim::econ::TokenAmount;
 use cid::Cid;
 use clap::Subcommand;
 use serde_tuple::{self, Deserialize_tuple, Serialize_tuple};
-
-use super::handle_rpc_err;
 
 #[derive(Serialize_tuple, Deserialize_tuple, Clone, Debug)]
 struct VestingSchedule {
@@ -35,15 +32,10 @@ pub enum StateCommands {
 }
 
 impl StateCommands {
-    pub async fn run(self, rpc_token: Option<String>) -> anyhow::Result<()> {
+    pub async fn run(self, api: ApiInfo) -> anyhow::Result<()> {
         match self {
             Self::Fetch { root, save_to_file } => {
-                println!(
-                    "{}",
-                    state_fetch_root((LotusJson(root), save_to_file), &rpc_token)
-                        .await
-                        .map_err(handle_rpc_err)?
-                );
+                println!("{}", api.state_fetch_root(root, save_to_file).await?);
             }
         }
         Ok(())