+++ src/cli_shared/cli/client.rs
@@ -74,6 +74,8 @@ pub struct Client {
     pub token_exp: Duration,
     /// Display progress bars mode. Auto will display if TTY.
     pub show_progress_bars: ProgressBarVisibility,
+    /// Load actors from the bundle file (possibly generating it if it doesn't exist)
+    pub load_actors: bool,
 }
 
 impl Default for Client {
@@ -98,6 +100,7 @@ impl Default for Client {
             rpc_address: SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), DEFAULT_PORT),
             token_exp: Duration::seconds(5184000), // 60 Days = 5184000 Seconds
             show_progress_bars: Default::default(),
+            load_actors: true,
         }
     }
 }