+++ src/cli_shared/cli/mod.rs
@@ -147,9 +147,15 @@ pub struct CliOpts {
     /// Disable the automatic database garbage collection.
     #[arg(long)]
     pub no_gc: bool,
+    /// In stateless mode, forest connects to the P2P network but does not sync to HEAD.
+    #[arg(long)]
+    pub stateless: bool,
     /// Check your command-line options and configuration file if one is used
     #[arg(long)]
     pub dry_run: bool,
+    /// Skip loading actors from the actors bundle.
+    #[arg(long)]
+    pub skip_load_actors: bool,
 }
 
 impl CliOpts {
@@ -247,6 +253,8 @@ impl CliOpts {
         if let Some(encrypt_keystore) = self.encrypt_keystore {
             cfg.client.encrypt_keystore = encrypt_keystore;
         }
+
+        cfg.client.load_actors = !self.skip_load_actors;
 
         Ok((cfg, path))
     }