+++ src/daemon/mod.rs
@@ -23,7 +23,7 @@ use crate::genesis::{get_network_name_from_genesis, read_genesis_header};
 use crate::key_management::{
     KeyStore, KeyStoreConfig, ENCRYPTED_KEYSTORE_NAME, FOREST_KEYSTORE_PHRASE_ENV,
 };
-use crate::libp2p::{Libp2pConfig, Libp2pService, PeerId, PeerManager};
+use crate::libp2p::{Libp2pConfig, Libp2pService, PeerManager};
 use crate::message_pool::{MessagePool, MpoolConfig, MpoolRpcProvider};
 use crate::rpc::start_rpc;
 use crate::rpc_api::data_types::RPCState;
@@ -150,11 +150,6 @@ pub(super) async fn start(
     let path: PathBuf = config.client.data_dir.join("libp2p");
     let net_keypair = crate::libp2p::keypair::get_or_create_keypair(&path)?;
 
-    // Hint at the multihash which has to go in the `/p2p/<multihash>` part of the
-    // peer's multiaddress. Useful if others want to use this node to bootstrap
-    // from.
-    info!("PeerId: {}", PeerId::from(net_keypair.public()));
-
     let mut keystore = load_or_create_keystore(&config).await?;
 
     if keystore.get(JWT_IDENTIFIER).is_err() {
@@ -181,7 +176,10 @@ pub(super) async fn start(
     )?)));
     let forest_car_db_dir = db_root_dir.join("car_db");
     load_all_forest_cars(&db, &forest_car_db_dir)?;
-    load_actor_bundles(&db).await?;
+
+    if config.client.load_actors {
+        load_actor_bundles(&db).await?;
+    }
 
     let mut services = JoinSet::new();
 
@@ -297,7 +295,8 @@ pub(super) async fn start(
         net_keypair,
         &network_name,
         genesis_cid,
-    )?;
+    )
+    .await?;
 
     let network_rx = p2p_service.network_receiver();
     let network_send = p2p_service.network_sender();
@@ -326,6 +325,7 @@ pub(super) async fn start(
         tipset_sink,
         tipset_stream,
         config.sync.clone(),
+        opts.stateless,
     )?;
     let bad_blocks = chain_muxer.bad_blocks_cloned();
     let sync_state = chain_muxer.sync_state_cloned();
@@ -388,7 +388,7 @@ pub(super) async fn start(
 
     // Sets the latest snapshot if needed for downloading later
     let mut config = config;
-    if config.client.snapshot_path.is_none() {
+    if config.client.snapshot_path.is_none() && !opts.stateless {
         set_snapshot_path_if_needed(
             &mut config,
             epoch,