+++ src/lotus_json/mod.rs
@@ -358,7 +358,7 @@ impl<T> Stringify<T> {
 macro_rules! lotus_json_with_self {
     ($($domain_ty:ty),* $(,)?) => {
         $(
-            impl HasLotusJson for $domain_ty {
+            impl $crate::lotus_json::HasLotusJson for $domain_ty {
                 type LotusJson = Self;
                 fn snapshots() -> Vec<(serde_json::Value, Self)> {
                     unimplemented!("tests are trivial for HasLotusJson<LotusJson = Self>")
@@ -373,6 +373,7 @@ macro_rules! lotus_json_with_self {
         )*
     }
 }
+pub(crate) use lotus_json_with_self;
 
 lotus_json_with_self!(
     u32,
@@ -381,4 +382,82 @@ lotus_json_with_self!(
     String,
     chrono::DateTime<chrono::Utc>,
     serde_json::Value,
+    (),
+    std::path::PathBuf,
+    bool,
 );
+
+impl<T: HasLotusJson> HasLotusJson for (T,) {
+    type LotusJson = (T::LotusJson,);
+    fn snapshots() -> Vec<(serde_json::Value, Self)> {
+        unimplemented!("tests are trivial for HasLotusJson<LotusJson = Self>")
+    }
+    fn into_lotus_json(self) -> Self::LotusJson {
+        (self.0.into_lotus_json(),)
+    }
+    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
+        (HasLotusJson::from_lotus_json(lotus_json.0),)
+    }
+}
+
+impl<A: HasLotusJson, B: HasLotusJson> HasLotusJson for (A, B) {
+    type LotusJson = (A::LotusJson, B::LotusJson);
+    fn snapshots() -> Vec<(serde_json::Value, Self)> {
+        unimplemented!("tests are trivial for HasLotusJson<LotusJson = Self>")
+    }
+    fn into_lotus_json(self) -> Self::LotusJson {
+        (self.0.into_lotus_json(), self.1.into_lotus_json())
+    }
+    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
+        (
+            HasLotusJson::from_lotus_json(lotus_json.0),
+            HasLotusJson::from_lotus_json(lotus_json.1),
+        )
+    }
+}
+
+impl<A: HasLotusJson, B: HasLotusJson, C: HasLotusJson> HasLotusJson for (A, B, C) {
+    type LotusJson = (A::LotusJson, B::LotusJson, C::LotusJson);
+    fn snapshots() -> Vec<(serde_json::Value, Self)> {
+        unimplemented!("tests are trivial for HasLotusJson<LotusJson = Self>")
+    }
+    fn into_lotus_json(self) -> Self::LotusJson {
+        (
+            self.0.into_lotus_json(),
+            self.1.into_lotus_json(),
+            self.2.into_lotus_json(),
+        )
+    }
+    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
+        (
+            HasLotusJson::from_lotus_json(lotus_json.0),
+            HasLotusJson::from_lotus_json(lotus_json.1),
+            HasLotusJson::from_lotus_json(lotus_json.2),
+        )
+    }
+}
+
+impl<A: HasLotusJson, B: HasLotusJson, C: HasLotusJson, D: HasLotusJson> HasLotusJson
+    for (A, B, C, D)
+{
+    type LotusJson = (A::LotusJson, B::LotusJson, C::LotusJson, D::LotusJson);
+    fn snapshots() -> Vec<(serde_json::Value, Self)> {
+        unimplemented!("tests are trivial for HasLotusJson<LotusJson = Self>")
+    }
+    fn into_lotus_json(self) -> Self::LotusJson {
+        (
+            self.0.into_lotus_json(),
+            self.1.into_lotus_json(),
+            self.2.into_lotus_json(),
+            self.3.into_lotus_json(),
+        )
+    }
+    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
+        (
+            HasLotusJson::from_lotus_json(lotus_json.0),
+            HasLotusJson::from_lotus_json(lotus_json.1),
+            HasLotusJson::from_lotus_json(lotus_json.2),
+            HasLotusJson::from_lotus_json(lotus_json.3),
+        )
+    }
+}