+++ src/rpc_client/db_ops.rs
@@ -2,10 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 use crate::rpc_api::db_api::*;
-use jsonrpc_v2::Error;
 
-use crate::rpc_client::call;
+use super::{ApiInfo, JsonRpcError, RpcRequest};
 
-pub async fn db_gc((): DBGCParams, auth_token: &Option<String>) -> Result<DBGCResult, Error> {
-    call(DB_GC, (), auth_token).await
+impl ApiInfo {
+    pub async fn db_gc(&self) -> Result<(), JsonRpcError> {
+        self.call(Self::db_gc_req()).await
+    }
+
+    pub fn db_gc_req() -> RpcRequest<()> {
+        RpcRequest::new(DB_GC, ())
+    }
 }